requirements.txt
requests_unixsocket
requests
python-dotenv
